feat: AddButton 삭제, test 페이지에 샘플 작성
